Fix bugs in 19 port

+++ src/main/java/me/axialeaa/colorfulminihud/ColorfulLines.java
@@ -484,21 +484,19 @@ public class ColorfulLines
         var("dayFrom0", day),
         var("dayFrom1", day + 1),
         var("hour", "02d", (dayTicks / 1000 + 6) % 24),
-        var("min", "02d", (dayTicks * 0.06) % 60),
-        var("sec", "02d", (dayTicks * 3.6) % 60)
+        var("min", "02d", (int)(dayTicks * 0.06) % 60), // These casts are necessary
+        var("sec", "02d", (int)(dayTicks * 3.6) % 60)
         //#if MC >= 11900
-        //$$ ,
-        //$$ var("moon", switch ((int) (day % 8)) {
-        //$$  case 0 -> "Full moon";
-        //$$  case 1 -> "Waning gibbous";
-        //$$  case 2 -> "Last quarter";
-        //$$  case 3 -> "Waning crescent";
-        //$$  case 4 -> "New moon";
-        //$$  case 5 -> "Waxing crescent";
-        //$$  case 6 -> "First quarter";
-        //$$  case 7 -> "Waxing gibbous";
-        //$$  default -> "Invalid";
-        //$$})
+        //$$ , var("moon", switch((int)(day % 8)){
+        //$$   case 0 -> "Full moon";
+        //$$   case 1 -> "Waning gibbous";
+        //$$   case 2 -> "Last quarter";
+        //$$   case 3 -> "Waning crescent";
+        //$$   case 4 -> "New moon";
+        //$$   case 5 -> "Waxing crescent";
+        //$$   case 6 -> "First quarter";
+        //$$   case 7 -> "Waxing gibbous";
+        //$$   default -> "Invalid";})
         //#endif
       ));
     }),
